backend/pyproject.toml
[project]
name = "medication-price-chatbot-backend"
version = "1.0.0"
description = "Backend API for medication price comparison chatbot"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.2",
    "tavily-python>=0.3.3",
    "openai>=1.3.7",
    "anthropic>=0.7.8",
    "fastmcp>=0.1.0",
    "mcp>=0.5.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "main.py",
    "*.py",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
]
